+++ src/components/hoc/withNavigate.test.jsx
@@ -1,4 +1,5 @@
 import { render } from "utils/test/utils";
+import { expect } from "vitest";
 import { withNavigate } from "./withNavigate";
 
 vi.mock("react-router");
@@ -13,7 +14,7 @@ describe("Test withNavigate", () => {
     expect(mockComponent).toHaveBeenCalled();
     expect(mockComponent).toHaveBeenCalledWith(
       { navigate: undefined }, // cause mock router
-      expect.anything(),
+      undefined,
     );
   });
   test("should be found props passed", () => {
@@ -26,7 +27,7 @@ describe("Test withNavigate", () => {
     expect(mockComponent).toHaveBeenCalled();
     expect(mockComponent).toHaveBeenCalledWith(
       { navigate: undefined, dummyProps: dummyPropsValue }, // cause mock router
-      expect.anything(),
+      undefined,
     );
   });
 });